+++ CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGELOG
   #68.
 * In order to improve performance when lookups are done from multiple
   threads, a use of `synchronized` has been removed. GitHub #65 & #69.
+* `jackson-databind` has been upgraded to 2.11.0.
 
 1.3.1 (2020-03-03)
 ------------------